pyproject.toml
[project]
name = "postgres-mcp"
version = "0.1.0"
description = "PostgreSQL MCP Server with Azure Database support - enables AI assistants to interact with PostgreSQL databases through the Model Context Protocol"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.29.0",
    "mcp>=0.9.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "streamlit>=1.28.0",
    "azure-identity>=1.15.0",
    "azure-keyvault-secrets>=4.7.0",
]
